+++ apworld/brotato/__init__.py
@@ -1,6 +1,7 @@
 import logging
+from collections import Counter
 from dataclasses import asdict
-from typing import Any, ClassVar, Dict, List, Literal, Sequence, Set, Tuple, Union
+from typing import Any, ClassVar, Dict, List, Literal, Set, Tuple, Union
 
 from BaseClasses import MultiWorld, Region, Tutorial
 from worlds.AutoWorld import WebWorld, World
@@ -12,23 +13,25 @@ from .constants import (
     CRATE_DROP_GROUP_REGION_TEMPLATE,
     CRATE_DROP_LOCATION_TEMPLATE,
     DEFAULT_CHARACTERS,
+    DEFAULT_ITEM_WEIGHTS,
     LEGENDARY_CRATE_DROP_GROUP_REGION_TEMPLATE,
     LEGENDARY_CRATE_DROP_LOCATION_TEMPLATE,
     MAX_SHOP_SLOTS,
     NUM_WAVES,
     RUN_COMPLETE_LOCATION_TEMPLATE,
     WAVE_COMPLETE_LOCATION_TEMPLATE,
+    ItemRarity,
 )
-from .items import (
-    BrotatoItem,
-    ItemName,
-    filler_items,
-    item_name_groups,
-    item_name_to_id,
-    item_table,
+from .items import BrotatoItem, ItemName, filler_items, item_name_groups, item_name_to_id, item_table
+from .locations import BrotatoLocation, BrotatoLocationBase, location_name_groups, location_name_to_id, location_table
+from .options import (
+    BrotatoOptions,
+    CommonItemWeight,
+    ItemWeights,
+    LegendaryItemWeight,
+    RareItemWeight,
+    UncommonItemWeight,
 )
-from .locations import BrotatoLocation, location_name_groups, location_name_to_id, location_table
-from .options import BrotatoOptions
 from .rules import create_has_character_rule, create_has_run_wins_rule, legendary_loot_crate_item_rule
 
 logger = logging.getLogger("Brotato")
@@ -46,7 +49,7 @@ class BrotatoWeb(WebWorld):
             ["RampagingHippy"],
         )
     ]
-    theme = "dirt"
+    theme: str = "dirt"
 
 
 class BrotatoWorld(World):
@@ -72,25 +75,37 @@ class BrotatoWorld(World):
     location_name_to_id: ClassVar[Dict[str, int]] = location_name_to_id
     location_name_groups: ClassVar[Dict[str, Set[str]]] = location_name_groups
 
-    waves_with_checks: Sequence[int]
+    waves_with_checks: List[int]
     """Which waves will count as locations.
 
     Calculated from player options in generate_early.
     """
 
-    common_loot_crate_groups: Sequence[BrotatoLootCrateGroup]
+    wave_per_game_item: Dict[int, List[int]]
+    """The wave to use to generate each Brotato item received, by rarity. Stored as slot data.
+
+    Brotato items are generated from a pool determined by the rarity (or tier) and the wave the item was found/bought.
+    We want to emulate this behavior with the items we create here. When we generate the items to match the common loot
+    crate drop locations, we also assign a wave to each item. When the client receives the next item for a certain
+    rarity, it will lookup the next entry in the list for the rarity and use that as the wave when generating the
+    values.
+
+    We attempt to equally distribute the items over the 20 waves in a normal run, with a bias towards lower numbers.
+    """
+
+    common_loot_crate_groups: List[BrotatoLootCrateGroup]
     """Information about each common loot crate group, i.e. how many crates it has and how many wins it needs.
 
     Calculated from player options in generate_early.
     """
 
-    legendary_loot_crate_groups: Sequence[BrotatoLootCrateGroup]
+    legendary_loot_crate_groups: List[BrotatoLootCrateGroup]
     """Information about each legendary loot crate group, i.e. how many crates it has and how many wins it needs.
 
     Calculated from player options in generate_early.
     """
 
-    def __init__(self, world: MultiWorld, player: int):
+    def __init__(self, world: MultiWorld, player: int) -> None:
         super().__init__(world, player)
 
     def create_item(self, name: Union[str, ItemName]) -> BrotatoItem:
@@ -98,7 +113,7 @@ class BrotatoWorld(World):
             name = name.value
         return item_table[self.item_name_to_id[name]].to_item(self.player)
 
-    def generate_early(self):
+    def generate_early(self) -> None:
         waves_per_drop = self.options.waves_per_drop.value
         # Ignore 0 value, but choosing a different start gives the wrong wave results
         self.waves_with_checks = list(range(0, NUM_WAVES + 1, waves_per_drop))[1:]
@@ -121,7 +136,7 @@ class BrotatoWorld(World):
             num_starting_characters = self.options.num_starting_characters.value
             self._starting_characters = self.random.sample(CHARACTERS, num_starting_characters)
 
-    def set_rules(self):
+    def set_rules(self) -> None:
         num_required_victories = self.options.num_victories.value
         self.multiworld.completion_condition[self.player] = create_has_run_wins_rule(
             self.player, num_required_victories
@@ -143,9 +158,16 @@ class BrotatoWorld(World):
             # crate_drop_region.connect(character_region, f"Exit drop crates for {character}", rule=has_character_rule)
             # character_regions.append(character_region)
 
-        self.multiworld.regions.extend([menu_region, *loot_crate_regions, *legendary_crate_regions, *character_regions])
+        self.multiworld.regions.extend(
+            [
+                menu_region,
+                *loot_crate_regions,
+                *legendary_crate_regions,
+                *character_regions,
+            ]
+        )
 
-    def create_items(self):
+    def create_items(self) -> None:
         item_names: List[Union[ItemName, str]] = []
 
         for c in self._starting_characters:
@@ -153,11 +175,10 @@ class BrotatoWorld(World):
 
         item_names += [c for c in item_name_groups["Characters"] if c not in self._starting_characters]
 
-        # Add an item to receive for each crate drop location, as backfill
-        num_common_crate_drops = self.options.num_common_crate_drops.value
-        for _ in range(num_common_crate_drops):
-            # TODO: Can be any item rarity, but need to choose a ratio. Check wiki for rates?
-            item_names.append(ItemName.COMMON_ITEM)
+        # Add an item to receive for each common crate drop location. Try to match the distribution of
+        # common/uncommon/rare/legendary items as we can infer from the game itself.
+        common_loot_crate_items: List[ItemName] = self._create_common_loot_crate_items()
+        item_names.extend(common_loot_crate_items)
 
         num_legendary_crate_drops = self.options.num_legendary_crate_drops.value
         for _ in range(num_legendary_crate_drops):
@@ -185,7 +206,7 @@ class BrotatoWorld(World):
         itempool = [self.create_item(item_name) for item_name in item_names]
 
         total_locations = (
-            num_common_crate_drops + num_legendary_crate_drops + (len(self.waves_with_checks) * len(CHARACTERS))
+            len(common_loot_crate_items) + num_legendary_crate_drops + (len(self.waves_with_checks) * len(CHARACTERS))
         )
         num_filler_items = total_locations - len(itempool)
         itempool += [self.create_filler() for _ in range(num_filler_items)]
@@ -194,13 +215,13 @@ class BrotatoWorld(World):
 
         # Place "Run Won" items at the Run Win event locations
         for loc in self.location_name_groups["Run Win Specific Character"]:
-            item = self.create_item(ItemName.RUN_COMPLETE)
+            item: BrotatoItem = self.create_item(ItemName.RUN_COMPLETE)
             self.multiworld.get_location(loc, self.player).place_locked_item(item)
 
-    def generate_basic(self):
+    def generate_basic(self) -> None:
         pass
 
-    def get_filler_item_name(self):
+    def get_filler_item_name(self) -> str:
         return self.random.choice(self._filler_items)
 
     def fill_slot_data(self) -> Dict[str, Any]:
@@ -214,14 +235,17 @@ class BrotatoWorld(World):
             "num_legendary_crate_locations": self.options.num_legendary_crate_drops.value,
             "num_legendary_crate_drops_per_check": self.options.num_legendary_crate_drops_per_check.value,
             "legendary_crate_drop_groups": [asdict(g) for g in self.legendary_loot_crate_groups],
+            "wave_per_game_item": self.wave_per_game_item,
         }
 
     def _create_character_region(self, parent_region: Region, character: str) -> Region:
-        character_region = Region(f"In-Game ({character})", self.player, self.multiworld)
-        character_run_won_location = location_table[RUN_COMPLETE_LOCATION_TEMPLATE.format(char=character)]
+        character_region: Region = Region(f"In-Game ({character})", self.player, self.multiworld)
+        character_run_won_location: BrotatoLocationBase = location_table[
+            RUN_COMPLETE_LOCATION_TEMPLATE.format(char=character)
+        ]
         character_region.locations.append(character_run_won_location.to_location(self.player, parent=character_region))
 
-        character_wave_drop_location_names = [
+        character_wave_drop_location_names: List[str] = [
             WAVE_COMPLETE_LOCATION_TEMPLATE.format(wave=w, char=character) for w in self.waves_with_checks
         ]
         character_region.locations.extend(
@@ -272,3 +296,70 @@ class BrotatoWorld(World):
             regions.append(group_region)
 
         return regions
+
+    def _create_common_loot_crate_items(self) -> List[ItemName]:
+        """Create a list of items corresponding to the common loot crate locations.
+
+        This is intended ot be called by `create_items`, but it's split out because of its side effect (see below), and
+        it's sort of involved.
+
+        Creates a Brotato Common/Uncommon/Rare/Legendary item for each loot crate location, usign the weights defined
+        in the use options to randomly determine how many of each tier to create.
+
+        This also has a side effect: it instantiates the `wave_per_game_item` field which is used to populate a slot
+        data entry with the same name. This defines the wave to use when determining what item to create client-side. We
+        do this here since we have the information readily available.
+        """
+        weights: Tuple[int, int, int, int]
+        if self.options.item_weight_mode.value == ItemWeights.option_default:
+            weights = DEFAULT_ITEM_WEIGHTS
+        elif self.options.item_weight_mode.value == ItemWeights.option_chaos:
+            # Ask each weight class for their bounds separately in case we ever make them different.
+            weights = tuple(
+                self.random.randint(weight.range_start, weight.range_end)
+                for weight in [CommonItemWeight, UncommonItemWeight, RareItemWeight, LegendaryItemWeight]
+            )  # type: ignore
+        elif self.options.item_weight_mode.value == ItemWeights.option_custom:
+            weights = (
+                self.options.common_item_weight.value,
+                self.options.uncommon_item_weight.value,
+                self.options.rare_item_weight.value,
+                self.options.legendary_item_weight.value,
+            )
+        else:
+            raise ValueError(f"Unsupported item_weight_mode {self.options.item_weight_mode.value}.")
+
+        item_names_to_rarity = {
+            ItemName.COMMON_ITEM: ItemRarity.COMMON,
+            ItemName.UNCOMMON_ITEM: ItemRarity.UNCOMMON,
+            ItemName.RARE_ITEM: ItemRarity.RARE,
+            ItemName.LEGENDARY_ITEM: ItemRarity.LEGENDARY,
+        }
+        items = self.random.choices(
+            list(item_names_to_rarity.keys()),
+            weights=weights,
+            k=self.options.num_common_crate_drops.value,
+        )
+
+        # Create the wave each item should be generated with. In each rarity, increment the wave by one for each item,
+        # looping over at 20 (the max number of waves in a run), then sort the result so we have an even distribution of
+        # waves in increasing order.
+        item_counts = Counter(items)
+
+        # Include the legendary items added from legendary crate drop checks as well
+        item_counts[ItemName.LEGENDARY_ITEM] = (
+            item_counts.get(ItemName.LEGENDARY_ITEM, 0) + self.options.num_legendary_crate_drops.value
+        )
+
+        def generate_waves_per_item(num_items: int) -> List[int]:
+            # Evenly distribute the items over 20 waves, then sort so items received are generated with steadily
+            # increasing waves (aka they got steadily stronger).
+            return sorted((i % 20) + 1 for i in range(num_items))
+
+        # Use a default of 0 in case no items of a tier were created for whatever reason.
+        self.wave_per_game_item: Dict[int, List[int]] = {
+            rarity.value: generate_waves_per_item(item_counts.get(name, 0))
+            for name, rarity in item_names_to_rarity.items()
+        }
+
+        return items