+++ apworld/brotato/_loot_crate_groups.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass
-from typing import List, Tuple
+from typing import List
 
 
 @dataclass(frozen=True)
@@ -9,7 +9,7 @@ class BrotatoLootCrateGroup:
     wins_to_unlock: int
 
 
-def build_loot_crate_groups(num_crates: int, num_groups: int, num_victories: int) -> Tuple[BrotatoLootCrateGroup, ...]:
+def build_loot_crate_groups(num_crates: int, num_groups: int, num_victories: int) -> List[BrotatoLootCrateGroup]:
     # If the options specify more crate drop groups than number of required wins, clamp to the number of wins. This
     # makes the math simpler and ensures all items are accessible by go mode. Someone probably wants the option to have
     # items after completing their goal, but we're going to pretend they don't exist until they ask.
@@ -35,9 +35,13 @@ def build_loot_crate_groups(num_crates: int, num_groups: int, num_victories: int
         crates_allocated += crates_in_group
 
         loot_crate_groups.append(
-            BrotatoLootCrateGroup(index=group_count, num_crates=crates_in_group, wins_to_unlock=wins_to_unlock_group)
+            BrotatoLootCrateGroup(
+                index=group_count,
+                num_crates=crates_in_group,
+                wins_to_unlock=wins_to_unlock_group,
+            )
         )
         # Set this for the next group now. This is the easiest way to ensure group 1 requires 0 victories.
         wins_to_unlock_group = min(wins_to_unlock_group + num_wins_to_unlock_group, num_victories)
 
-    return tuple(loot_crate_groups)
+    return loot_crate_groups